% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_fragment_custom.R
\name{html_fragment_custom}
\alias{html_fragment_custom}
\title{Custom HTML fragment output format with syntax highlighting}
\usage{
html_fragment_custom(
  number_sections = FALSE,
  section_divs = TRUE,
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  df_print = "default",
  mathjax = TRUE,
  includes = NULL,
  keep_md = FALSE,
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)
}
\description{
Custom HTML fragment output format with syntax highlighting
}
